BusinessATM
For this project you are going to be creating a Java application that will allow business users to access time critical banking on-the-go. Users will be required to login to the application upon start-up and upon successful login will be directed to a menu where they will be able to manage their accounts with the bank or even create new accounts if they need to.
Users should be able to manage any existing accounts they have.
Users should be able to create new accounts if they don’t already have one of each type.
If a user does not have any accounts they should be asked if they would like to create a new one.
Restrictions will be placed on all accounts that require two signatories.
Small business, community and client  accounts can have overdraft facilities of £1,000, £2,500 and £1500 respectively.
Users can create Small Business, Community and Client Accounts
Users should be able to Check the balance  of their accounts as well as make withdrawals, deposits and transfers to other accounts that they own.